\name{configLimix}
\alias{configLimix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Configure limix
}
\description{
If limix isn't in the python path, a path to the limix package can be specified with this function.
}
\usage{
configLimix(limix_path)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{limix_path}{
A string to the limix path.
}
}
