\name{data_mESCquartz}
\alias{data_mESCquartz}
\docType{data}
\title{
Gene expression of 35 single mouse embryonic stem cells.
}
\description{
The gene expression (FPKM values) of 36807 genes of 35 mouse embryonic stem cells at differnt stages of the cell cycle, published in Sasagawa et al 2013.
}
\usage{data("data_mESCquartz")}
\format{
  A data frame containing the gene expression of 36807 genes of 35 cells . 
 
}

\source{
Sasagawa, Y., Nikaido, I., Hayashi, T., Danno, H., Uno, K. D., Imai, T., and Ueda, H. R. (2013). Quartz-Seq: a highly reproducible and sensitive single-cell RNA sequencing method, reveals non-genetic gene-expression heterogeneity. \emph{Genome Biol}, \bold{14}(4), R31.
}

\examples{
data(data_mESCquartz)

}
\keyword{datasets}
