\name{data_Tcells}
\alias{data_Tcells}
\docType{data}
\title{
Gene expression of 81 T-cells.
}
\description{
The raw read counts for 38385 genes (including ERCC spike-ins and endogenous genes) of 81 T-cells as described in Mahata et al.

}
\usage{data("data_Tcells")}
\format{
  A data frame with the expression of 38385 genes of 81 cells.

}

\source{
Mahata, B., Zhang, X., Kolodziejczyk, A. A., Proserpio, V., Haim-Vilmovsky, L., Taylor, A. E., ... and Teichmann, S. A. (2014). Single-cell RNA sequencing reveals T helper cells synthesizing steroids de novo to contribute to immune homeostasis. \emph{Cell reports}, \bold{7}(4), 1130-1142.
}
\references{
Buettner F, Natarajan KN, Casale FP, Proserpio V, Scialdone A, Theis FJ, Teichmann SA, Marioni JC and Stegle O, 2015. Computational analysis of cell-to-cell heterogeneity in single-cell RNA-Sequencing data reveals hidden subpopulation of cells, \emph{Nature Biotechnology}, doi: 10.1038/nbt.3102.
}
\examples{
data(data_Tcells)

}
\keyword{datasets}
